requirements.txt
loguru==0.7.2
bittensor==7.1.2
python-dotenv==1.0.1
pydantic-settings==2.6.0
wandb==0.18.5
numpy==2.0.2
pandas==2.2.3
websockets

